Expose switch ports
Ports of switches are now a full resource that is
identified by the switch that contains it together
with the resource id of the port.
Ports are exposed at
switches/<switch_id>/ports/<port_id> and support
GET, PUT, PATCH verbs. For POST, DELETE (adding
and removing ports) it is necessary to add / delete
the port models of the switches switch model.